fix(restart): improve the restart script..no default buildout

+++ restart_wrapper.sh
@@ -2,15 +2,18 @@
 BASEDIR=$(dirname "$0")
 cd $BASEDIR
 
-# step 2: force get latest and greatest code
-git reset --hard HEAD
-git pull
+LATEST=${1:-"NOPE"}
+if [ $LATEST == "YES" ]; then
+  # step 2: force get latest and greatest code
+  git reset --hard HEAD
+  git pull
 
-# step 3: remove (potentially) outdated dependencies and buildout to get latest / greatest
-RM_DEPS="rm -rf *egg*/ott.utils* nohup.out *~"
-echo $RM_DEPS
-eval $RM_DEPS
-buildout
+  # step 3: remove (potentially) outdated dependencies and buildout to get latest / greatest
+  RM_DEPS="rm -rf *egg*/ott.utils* nohup.out *~"
+  echo $RM_DEPS
+  eval $RM_DEPS
+  buildout
+fi
 
 # step 4: stop & restart wrapper
 pkill -9 -f -c local_pelias
@@ -18,5 +21,7 @@ cmd="bin/pserve config/local_pelias.ini"
 eval "${cmd}" &>/dev/null & disown;
 
 # step 5: really done hopefully...
-echo 
+echo
+ps -ef | grep pserve | grep -v grep
 echo "DONE"
+echo
